test: split feature page tests by page
